Moving passport and svelte code to common lib.
