--- lib/service/config/default.json
@@ -1,16 +1,0 @@
-{
-	"logging": {
-		"appenders": [{"type": "console"}],
-		"levels": {
-			"[all]": "DEBUG"
-		}
-	},
-	"session": {
-		"secret": "dcOVe-ZW3ul77l23GiQSNbTJtMRio87G2yUOUAk_otcbL3uywfyLMZ9NBmDMuuOt"
-	},
-	"authentication": {
-		"serviceUrl": "https://identity.dev.cshr.digital:9443",
-		"serviceAdmin": "admin@cslearning.gov.uk",
-		"servicePassword": "admin"
-	}
-}
