--- lib/service/config/production.json
@@ -1,7 +1,0 @@
-{
-  "logging": {
-    "levels": {
-      "[all]": "DEBUG"
-    }
-  }
-}
