--- lib/service/config/test.json
@@ -1,7 +1,0 @@
-{
-  "logging": {
-    "levels": {
-      "[all]": "DEBUG"
-    }
-  }
-}
