+++ lib/ui/svelte.d.ts
