+++ service/course-ui/package.json
@@ -23,7 +23,6 @@
 		"@types/session-file-store": "1.2.0"
 	},
 	"dependencies": {
-		"@types/request": "^2.0.13",
 		"compression": "1.7.1",
 		"config": "1.29.0",
 		"express": "4.16.2",
@@ -36,9 +35,7 @@
 		"node-fetch": "1.7.3",
 		"node-sass": "4.7.2",
 		"passport": "0.3.2",
-		"passport-saml": "0.32.1",
 		"request": "^2.83.0",
-		"session-file-store": "1.2.0",
-		"svelte": "1.49.1"
+		"session-file-store": "1.2.0"
 	}
 }
