+++ service/ui/package.json
@@ -30,13 +30,11 @@
 		"express-validator": "4.3.0",
 		"govuk-elements-sass": "3.1.2",
 		"govuk_frontend_toolkit": "7.2.0",
+		"lib": "file:../../lib",
 		"lusca": "1.5.2",
 		"node-fetch": "1.7.3",
 		"node-sass": "4.7.2",
 		"passport": "0.3.2",
-		"passport-saml": "0.32.1",
-		"session-file-store": "1.2.0",
-		"svelte": "1.49.1",
-		"lib": "file:../../lib"
+		"session-file-store": "1.2.0"
 	}
 }
