+++ service/ui/src/controllers/search.ts
@@ -1,5 +1,5 @@
 import {Request, Response} from 'express'
-import * as template from 'ui/template'
+import * as template from 'lib/ui/template'
 import * as catalog from 'lib/service/catalog'
 import * as elko from 'lib/service/elko'
