--- service/ui/src/env.ts
@@ -1,1 +1,0 @@
-export const PRODUCTION = process.env.NODE_ENV === 'production'
